% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acknowledgements.R
\name{ANSIS_Ackowledgement}
\alias{ANSIS_Ackowledgement}
\title{ANSIS Funding Ackowledgement}
\usage{
ANSIS_Ackowledgement()
}
\value{
character
}
\description{
ANSIS Funding Ackowledgement
}
\author{
Ross Searle
}
