% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{ANSIS_APIDocs}
\alias{ANSIS_APIDocs}
\title{Get ANSIS API Documentation}
\usage{
ANSIS_APIDocs()
}
\value{
file
}
\description{
Get ANSIS API Documentation
}
\details{
Opens up the documentation of the ANSIS API upon wich this package has been developed
}
\author{
Ross Searle
}
