% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{ANSIS_Demos}
\alias{ANSIS_Demos}
\title{Open a demo R Script}
\usage{
ANSIS_Demos()
}
\value{
R script
}
\description{
Open a demo R Script
}
\details{
Opens up a script in RStudio which demonstrates the functionality in the 'ANSR' package.
}
\author{
Ross Searle
}
