% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{ANSIS_OpenWebsite}
\alias{ANSIS_OpenWebsite}
\title{Opens the ANSIS Website}
\usage{
ANSIS_OpenWebsite()
}
\value{
url
}
\description{
Opens the ANSIS Website
}
\details{
Opens up the ANSIS Website
}
\author{
Ross Searle
}
