% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{ANSISOpenWebsite}
\alias{ANSISOpenWebsite}
\title{Opens the ANSIS Website}
\usage{
ANSISOpenWebsite()
}
\value{
url
}
\description{
Opens the ANSIS Website
}
\details{
Opens up the ANSIS Website
}
\author{
Ross Searle
}
