% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{apiAuthoriseMe}
\alias{apiAuthoriseMe}
\title{Authorise ANSIS Session}
\usage{
apiAuthoriseMe(username, password, DataStorePath)
}
\arguments{
\item{username}{ANSIS account username}

\item{password}{ANSIS account password}
}
\value{
logical
}
\description{
Authorise ANSIS Session
}
\details{
Before you can use this package you need to provide your ANSIS account credentials. If you do not already have a valid ANSIS user account please go to the ANSI website to creat an account.
}
\examples{
AuthoriseMe(username='Bob', password='123')
}
\author{
Ross Searle
}
