% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{apiCatalogueSummary}
\alias{apiCatalogueSummary}
\title{ANSIS Catalogue Summary}
\usage{
apiCatalogueSummary()
}
\value{
data frame
}
\description{
ANSIS Catalogue Summary
}
\details{
Lists the ANSIS data providers
}
\author{
Ross Searle
}
