% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{apiDeleteAllQueries}
\alias{apiDeleteAllQueries}
\title{Delete all ANSIS Queries}
\usage{
apiDeleteAllQueries()
}
\value{
ANSIS query status
}
\description{
Delete all ANSIS Queries
}
\details{
Deletes (stops) all of your ANSIS Data System queries
}
\author{
Ross Searle
}
