% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{apiDeleteQuery}
\alias{apiDeleteQuery}
\title{Delete an ANSIS Query}
\usage{
apiDeleteQuery(reqID)
}
\arguments{
\item{reqID}{Request ID}
}
\value{
ANSIS query status
}
\description{
Delete an ANSIS Query
}
\details{
Deletes (stops) a specific ANSIS Data System query
}
\author{
Ross Searle
}
