% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{apiDownloadQueryData}
\alias{apiDownloadQueryData}
\title{Download Query Data}
\usage{
apiDownloadQueryData(reqID, outDir = NULL)
}
\arguments{
\item{reqID}{Request ID}

\item{outDir}{A directory path that the data will be downloaded to.}
}
\value{
json ANSIS data
}
\description{
Download Query Data
}
\details{
Downloads the json query response files from the ANSIS server to your local machine. It also merges the single ANSIS response files into a single json file
}
\author{
Ross Searle
}
