% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{apiGenerateToken}
\alias{apiGenerateToken}
\title{Generate an Authorisation token}
\usage{
apiGenerateToken(user, pwd, verbose = F)
}
\arguments{
\item{user}{ANSIS account username}

\item{pwd}{ANSIS account password}
}
\value{
logical
}
\description{
Generate an Authorisation token
}
\details{
The ANSI API requires users to authenticate with a token.
}
\author{
Ross Searle
}
