% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{apiGetANSISData}
\alias{apiGetANSISData}
\title{Get ANSIS Data}
\usage{
apiGetANSISData(
  Name = NULL,
  Description = NULL,
  minx = minx,
  maxx = NULL,
  miny = NULL,
  maxy = NULL,
  soilProperty = NULL,
  propertyName = NULL,
  labCode = NULL,
  startYear = NULL,
  endYear = NULL,
  provider = NULL,
  sites = NULL,
  numCPUs = NULL
)
}
\arguments{
\item{Name}{A name for the query. Default is NULL.}

\item{Description}{A description of the query.}

\item{minx}{Western extent of the bounding box of the query. CRS WGS84. Default is NULL.}

\item{maxx}{Eastern extent of the bounding box of the query. CRS WGS84. Default is NULL.}

\item{miny}{Sothern extent of the bounding box of the query. CRS WGS84. Default is NULL.}

\item{maxy}{Northern extent of the bounding box of the query. CRS WGS84. Default is NULL.}

\item{soilProperty}{Soil property element to query on. This is the broadest ANSIS attribute grouping.}

\item{propertyName}{A specific form of the soil property.}

\item{labCode}{Green Book standard lab method code.}

\item{startYear}{Starting year of date range to return data for.}

\item{endYear}{Finishing year of date range to return data for.}

\item{provider}{The Data Provider to return data for.}

\item{sites}{A list of ANSIS Site IDs to return data for.}
}
\value{
ANSIS Data Object
}
\description{
Get ANSIS Data
}
\details{
This is the main function to use to retrieve data form the ANSIS Data System. It wraps a number of individual functions contained in this package to process a query of the ANSIS Data System from start to finish. THis function sends a query to the ANSIS Data System, retrieves the data and massages the response into and ANSIS Data Object.
}
\author{
Ross Searle
}
