% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{apiPropertyDefinitions}
\alias{apiPropertyDefinitions}
\title{ANSIS Property Definitions}
\usage{
apiPropertyDefinitions()
}
\value{
data frame
}
\description{
Returns a data frame of ANSIS property codes and their definitions
}
\examples{
head(apiPropertyDefinitions())
}
\author{
Ross Searle
}
