% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{apiProviderCatalogue}
\alias{apiProviderCatalogue}
\title{ANSIS Provider Catalogue}
\usage{
apiProviderCatalogue(poviderNames = NULL, includeProperties = F)
}
\value{
JSON list
}
\description{
Shows details for a specific ANSIS data provider
}
\details{
Returns the catalogue for the provider specified in the {ID}. Valid IDs are provided via the “name” field of the apiCatalogueSummary() function. 
Returns a list of ANSIS site numbers along with locations and a list of properties for each site.
}
\author{
Ross Searle
}
