% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{apiQueryStatus_All}
\alias{apiQueryStatus_All}
\title{Get the Statuts of all ANSIS Queries}
\usage{
apiQueryStatus_All(statusTypes = NULL)
}
\arguments{
\item{statusTypes}{If specified just statuses of the given types are returned.}
}
\value{
data frame
}
\description{
Get the Statuts of all ANSIS Queries
}
\details{
Gets the status of all of your ANSIS Data System queries.
}
\author{
Ross Searle
}
