% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{apiQueryStatus_Monitor}
\alias{apiQueryStatus_Monitor}
\title{Monitor an ANSIS Query Status}
\usage{
apiQueryStatus_Monitor(reqID)
}
\arguments{
\item{reqID}{get the status for this request ID}
}
\value{
ANSIS query status
}
\description{
Monitor an ANSIS Query Status
}
\details{
Monitors the status of an ANSIS query until it is finished
}
\author{
Ross Searle
}
