% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{apiQueryStatus_Single}
\alias{apiQueryStatus_Single}
\title{Get an ANSIS Query Status}
\usage{
apiQueryStatus_Single(reqID, verbose = F)
}
\arguments{
\item{reqID}{Request ID}
}
\value{
ANSIS query status
}
\description{
Get an ANSIS Query Status
}
\details{
Returns the status of a specific requiest ID.
}
\author{
Ross Searle
}
