% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{apiSendQuery}
\alias{apiSendQuery}
\title{Send a Query to the ANSIS API}
\usage{
apiSendQuery(
  minx = minx,
  maxx = NULL,
  miny = NULL,
  maxy = NULL,
  soilProperty = NULL,
  propertyName = NULL,
  labCode = NULL,
  startYear = NULL,
  endYear = NULL,
  provider = NULL,
  sites = NULL
)
}
\arguments{
\item{minx}{Western extent of the bounding box of the query. CRS WGS84. Default is NULL.}

\item{maxx}{Eastern extent of the bounding box of the query. CRS WGS84. Default is NULL.}

\item{miny}{Sothern extent of the bounding box of the query. CRS WGS84. Default is NULL.}

\item{maxy}{Northern extent of the bounding box of the query. CRS WGS84. Default is NULL.}

\item{soilProperty}{Soil property element to query on. This is the broadest ANSIS attribute grouping.}

\item{propertyName}{A specific form of the soil property.}

\item{labCode}{Green Book standard lab method code.}

\item{startYear}{Starting year of date range to return data for.}

\item{endYear}{Finishing year of date range to return data for.}

\item{provider}{The Data Provider to return data for.}

\item{sites}{A list of ANSIS Site IDs to return data for.}

\item{Name}{A name for the query. Default is NULL.}

\item{Description}{A description of the query.}
}
\value{
Request ID
}
\description{
Send a Query to the ANSIS API
}
\details{
Sends a query to the ANSIS Data System.
}
\author{
Ross Searle
}
