% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryCache.R
\name{cacheEmpty}
\alias{cacheEmpty}
\title{Empty the Local Data Cache}
\usage{
cacheEmpty(queryName = NULL)
}
\arguments{
\item{queryName}{If specified the function only deletes data for that specific query}
}
\value{
data frame
}
\description{
Deletes cached queries
}
\author{
Ross Searle
}
