% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryCache.R
\name{cacheRetrieveData}
\alias{cacheRetrieveData}
\title{Retrieve Data From Local Cache}
\usage{
cacheRetrieveData(qName)
}
\arguments{
\item{qName}{The name of the query to retrieve data for}
}
\value{
ANSIS Data Object - List
}
\description{
When you do a query on the ANSIS data system using the apiGetANSISData() funtion, the data returned from ANSIS is cached locally. Data is returned from the local cache if you send a query with exactly the same parameters again.
}
\author{
Ross Searle
}
