% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryCache.R
\name{cacheShowCachedQueries}
\alias{cacheShowCachedQueries}
\title{Show Cached Queries}
\usage{
cacheShowCachedQueries(queryName = NULL, verbose = T)
}
\arguments{
\item{queryName}{If specified the function only returns information for that specific query}

\item{verbose}{Show all information}
}
\value{
data frame
}
\description{
Shows information about cached queries
}
\author{
Ross Searle
}
