% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryCache.R
\name{cacheShowFileSizes}
\alias{cacheShowFileSizes}
\title{Show Cache File Sizes}
\usage{
cacheShowFileSizes(siteCount = F)
}
\arguments{
\item{siteCount}{If true this function will also show the number of sites returned by each query}
}
\value{
ANSIS Data Object - List
}
\description{
Shows a list of the files stored in the local data cahe and their file size
}
\author{
Ross Searle
}
