% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryCache.R
\name{cacheShowQueryNames}
\alias{cacheShowQueryNames}
\title{Show Local Data Cache QueryNames}
\usage{
cacheShowQueryNames()
}
\value{
list
}
\description{
A list of the query names currently stored in the local data cache
}
\author{
Ross Searle
}
