% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseANSISJson.R
\name{getAvailableProperties}
\alias{getAvailableProperties}
\title{Return all data for a site}
\usage{
getAvailableProperties(anisObject, propertyType = NULL)
}
\arguments{
\item{anisObject}{An ANSIS R object}

\item{propertyType}{Return data for either 'SiteVisit', 'Horizons' or 'Lab'}
}
\value{
dataframe
}
\description{
Return all data for a site
}
\details{
Return all of the soil properties names that are available in the ANSIS R object
}
\author{
Ross Searle
}
