% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{getDSMtable}
\alias{getDSMtable}
\title{Get a DSM Suitable Data Table}
\usage{
getDSMtable(
  Name = NULL,
  Description = NULL,
  minx,
  maxx,
  miny,
  maxy,
  soilProperty = NULL,
  propertyName = NULL,
  labCode = NULL,
  startYear = 1900,
  endYear = NULL
)
}
\arguments{
\item{Name}{ANSIS account username}

\item{Description}{ANSIS account password}
}
\value{
logical
}
\description{
Get a DSM Suitable Data Table
}
\details{
place holder
}
\author{
Ross Searle
}
