% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiHelpers.R
\name{getPropertiesInANSISResponse}
\alias{getPropertiesInANSISResponse}
\title{Get Properties in an ANSIS Response}
\usage{
getPropertiesInANSISResponse(ANSISObject, PropertyType = NULL)
}
\arguments{
\item{ANSISObject}{An ANSIS data object}

\item{PropertyType}{Used to filter the response on a property}
}
\value{
ANSIS query status
}
\description{
Get Properties in an ANSIS Response
}
\details{
Returns the status of a specific requiest ID.
}
\author{
Ross Searle
}
