% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseANSISJson.R
\name{getPropertiesSummary}
\alias{getPropertiesSummary}
\title{Return a summary of the number of soil properties}
\usage{
getPropertiesSummary(ansisObject)
}
\arguments{
\item{anisObject}{An ANSIS R object}
}
\value{
dataframe
}
\description{
Return a summary of the number of soil properties
}
\details{
Return a summary of the number of soil properties in an ANSIS R object as a dataframe
}
\author{
Ross Searle
}
