% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseANSISJson.R
\name{getSiteData}
\alias{getSiteData}
\title{Return all data for a site}
\usage{
getSiteData(ansisObject, siteID)
}
\arguments{
\item{siteID}{Site ID}

\item{anisObject}{An ANSIS R object}
}
\value{
dataframe
}
\description{
Return all data for a site
}
\details{
Return all data for a site from an ANSIS R object as a dataframe
}
\author{
Ross Searle
}
