% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseANSISJson.R
\name{getSiteLocations}
\alias{getSiteLocations}
\title{Return all site locations}
\usage{
getSiteLocations(ansisObject)
}
\arguments{
\item{anisObject}{An ANSIS R object}
}
\description{
Return all site locations
}
\details{
Return all site locations from an ANSIS R object as a dataframe
}
\author{
Ross Searle
}
