% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryCache.R
\name{makeQueryObject}
\alias{makeQueryObject}
\title{library(DBI)}
\usage{
makeQueryObject(
  Name = NULL,
  Description = NULL,
  minx = minx,
  maxx = NULL,
  miny = NULL,
  maxy = NULL,
  soilProperty = NULL,
  propertyName = NULL,
  labCode = NULL,
  startYear = NULL,
  endYear = NULL,
  provider = NULL,
  sites = NULL
)
}
\description{
library(DBI)
}
