% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseANSISJson.R
\name{makeWideTable}
\alias{makeWideTable}
\title{# getAO
#
#' Example Data
#' @export
getAO <- function(){
  ao <- ansisExampleObject
  return(ao)
}
Format the ANSIS response into a wide table}
\usage{
makeWideTable(ansisObject, propertyType = NULL, labcodes = NULL, decode = F)
}
\arguments{
\item{propertyType}{Return data for a specific list of soil properties}

\item{decode}{return either coded or decoded values (default=F)}

\item{anisObject}{An ANSIS R object}
}
\value{
dataframe
}
\description{
# getAO
#
#' Example Data
#' @export
getAO <- function(){
  ao <- ansisExampleObject
  return(ao)
}
Format the ANSIS response into a wide table
}
\details{
You need to specify a parameter value for either 'propertyType' or 'properties'. 
The data frame contains SiteID, location, depths and soil property fields
}
\author{
Ross Searle
}
