% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseANSISJson.R
\name{parseANSISJson}
\alias{parseANSISJson}
\title{Parses an ANSIS JSON response into an R ANSIS object}
\usage{
parseANSISJson(jsnFile)
}
\arguments{
\item{jsnFile}{The path to the ANSIS JSON file to parse}
}
\value{
data frame
}
\description{
Parses an ANSIS JSON response into an R ANSIS object
}
\details{
This function returns a spectra dataframe from the CSIS Database for a given Transformation ID. If a labCode is supplied only Spectra records with a related lab method are returned
}
\examples{
getTransformedSpectraAsDataFrame(TransformationID=2, labCode='4A1')
}
\author{
Ross Searle
}
