% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansisPropertCodes.R
\name{showAnsisPropertyCodes}
\alias{showAnsisPropertyCodes}
\title{Show ANSIS Property Code Mappings}
\usage{
showAnsisPropertyCodes(
  soilProperty = NULL,
  propertyName = NULL,
  labCode = NULL
)
}
\arguments{
\item{soilProperty}{The broadest grouping of soil properties. ie at the elemental level eg 'Nitrogen'.}

\item{propertyName}{A broad laboratory method name.}

\item{labCode}{A Green Book lab method code.}
}
\value{
data.frame
}
\description{
Show ANSIS Property Code Mappings
}
\details{
ANSIS has its own specific code to identify soil properties. The information in this table is used to map to these codes to allow querying of the ANSIS data repository
}
\author{
Ross Searle
}
