% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{apiProviderCatalogue}
\alias{apiProviderCatalogue}
\title{ANSIS Provider Catalogue}
\usage{
apiProviderCatalogue(
  poviderNames = NULL,
  includeProperties = F,
  outputFormat = "dataframe",
  ViewIt = F
)
}
\arguments{
\item{poviderNames}{A list of ANSIS Provider names.}

\item{includeProperties}{Include the list of ANSIS Soil properties associated with each site.}

\item{outputFormat}{Either a dataframe or JSON}

\item{ViewIt}{View the raw json output.}
}
\value{
dataframe
}
\description{
Shows details for specified ANSIS data providers
}
\details{
Returns the site data catalogues for the providers specified with the poviderNames list. Valid IDs are provided via the “name” field of the apiCatalogueSummary() function.
Returns a list of ANSIS site numbers along with locations and a list of properties for each site.
}
\author{
Ross Searle
}
