% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{getDSMtable}
\alias{getDSMtable}
\title{Get a DSM Suitable Data Table}
\usage{
getDSMtable(
  Name = NULL,
  Description = NULL,
  minx,
  maxx,
  miny,
  maxy,
  soilProperty = NULL,
  propertyName = NULL,
  labCode = NULL,
  startYear = 1900,
  endYear = NULL
)
}
\arguments{
\item{Name}{A name for the query. Default is NULL.}

\item{Description}{A description of the query.}

\item{minx}{Western extent of the bounding box of the query. CRS WGS84. Default is NULL.}

\item{maxx}{Eastern extent of the bounding box of the query. CRS WGS84. Default is NULL.}

\item{miny}{Sothern extent of the bounding box of the query. CRS WGS84. Default is NULL.}

\item{maxy}{Northern extent of the bounding box of the query. CRS WGS84. Default is NULL.}

\item{soilProperty}{Soil property element to query on. This is the broadest ANSIS attribute grouping.}

\item{propertyName}{A specific form of the soil property.}

\item{labCode}{Green Book standard lab method code.}

\item{startYear}{Starting year of date range to return data for.}

\item{endYear}{Finishing year of date range to return data for.}
}
\value{
logical
}
\description{
Get a DSM Suitable Data Table
}
\details{
This is a convenience function which returns data from An ANSIS query in a format for typical Digital Soil Mapping workflows. ie Site IDs, Locations, Dates, Depths and Soil Property values in a wide table format.
}
\author{
Ross Searle
}
