% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{getDSMtable}
\alias{getDSMtable}
\title{Get a DSM Suitable Data Table}
\usage{
getDSMtable(ansisObject = NULL, propertyCode = NULL)
}
\arguments{
\item{ansisObject}{an ANSIS data object.}

\item{propertyCode}{Soil property code.}
}
\value{
logical
}
\description{
Get a DSM Suitable Data Table
}
\details{
This is a convenience function which returns data from An ANSIS query in a format for typical Digital Soil Mapping workflows. ie Site IDs, Locations, Dates, Depths and Soil Property values in a wide table format.
}
\author{
Ross Searle
}
